+++ classes/prefix_WPgutenberg/blocks/posts/index.php
@@ -65,7 +65,11 @@ register_block_type(
       'postPopUpNav' => array(
         'type' => 'boolean',
         'default' => false
-      )
+      ),
+      'postSortNav' => array(
+        'type' => 'boolean',
+        'default' => false
+      ),
     ),
     // 'render_callback' => function($atts) {
     //   return SELF::WPgutenberg_blockRender_posts($atts);
@@ -98,42 +102,41 @@ function WPgutenberg_block_postsValue($value, $id){
             return '<figure>' . $img_url . '</figure>';
           endif;
         else:
-          return get_post_meta($id, $value, true);
+          $metaValue = get_post_meta($id, $value, true);
+          if(is_array($metaValue)):
+            foreach ($metaValue as $key => $value) {
+              return '<span>' . get_the_title($value) . '</span>';
+            }
+          else:
+            return $metaValue;
+          endif;
         endif;
       endif;
   }
 }
 
-// callback function
-function WPgutenberg_blockRender_posts($attr){
-  $output = '';
-  $css = '';
-  $inlinecss = '';
-  // add edvanced options (ID/CSS)
-  $id = array_key_exists('anchor', $attr) ? ' id="' . $attr['anchor'] . '"' : '';
-  $css .= array_key_exists('className', $attr) ? ' ' . $attr['className'] : '';
-  $css .= array_key_exists('align', $attr) ? ' align' . $attr['align'] : '';
-  // add swiper
-  $css .= array_key_exists('postSwiper', $attr) && $attr['postSwiper'] !== false ? ' gallery-swiper' : ' gallery-grid';
-  // add popup
-  $css .= array_key_exists('postPopUp', $attr) && $attr['postPopUp'] !== false ? ' add-popup' : '';
-  $css .= array_key_exists('postPopUpNav', $attr) && $attr['postPopUpNav'] !== false ? ' popup-preview' : '';
-  // reset spacing if only one column
-  $spacing = $attr['postColumnsSpace'];
-  // inline stylings
-  $columns = array_key_exists('postColumns', $attr) ? $attr['postColumns'] : 1;
-  $inlinecss .= '--postColumns: ' . $columns . ';';
-  $inlinecss .= ' --postColumnsSpace: ' . $spacing . 'px;';
+// post results
+function WPgutenberg_postresults_postssorting(array $attr, string $source = 'first_load'){
+  // sort direction
+  $postSortBy = in_array($attr['postSortBy'], array('title', 'date', 'menu_order')) ? $attr['postSortBy'] : 'meta_value';
   // add filter
   $filter = [];
   if(array_key_exists('postTaxonomyFilter', $attr)):
-    foreach ($attr['postTaxonomyFilter'] as $key => $value) {
-      $stringToArray = explode("-", $value);
-      if(array_key_exists($stringToArray[0], $filter)):
-        $filter[$stringToArray[0]] = [];
-      endif;
-      $filter[$stringToArray[0]][] = $stringToArray[1];
-    }
+    if($source == 'first_load'):
+      foreach ($attr['postTaxonomyFilter'] as $key => $value) {
+        $stringToArray = explode("-", $value);
+        // if(array_key_exists($stringToArray[0], $filter)):
+        //   $filter[$stringToArray[0]] = [];
+        // endif;
+        $filter[$stringToArray[0]][] = $stringToArray[1];
+      }
+    elseif($source == 'ajax'):
+      $thefilter = explode("__", $attr['postTaxonomyFilter']);
+      foreach ($thefilter as $key => $value) {
+        $stringToArray = explode("-", $value);
+        $filter[$stringToArray[0]][] = $stringToArray[1];
+      }
+    endif;
   endif;
   if(!empty($filter)):
     $term_array = array('relation' => $attr['postTaxonomyFilterRelation']);
@@ -166,9 +169,12 @@ function WPgutenberg_blockRender_posts($attr){
     'post_status'=>'publish',
     'posts_per_page'=> $attr['postSum'],
     'paged' => $paged,
-    'orderby' => $attr['postSortBy'],
+    'orderby' => $postSortBy,
     'order' => $attr['postSortDirection']
   );
+  if(!in_array($attr['postSortBy'], array('title', 'date', 'menu_order'))):
+    $queryArgs['meta_key'] = $attr['postSortBy'];
+  endif;
   if($attr['postType'] == 'attachment'):
     $queryArgs['post_status'] = 'inherit';
   endif;
@@ -180,84 +186,156 @@ function WPgutenberg_blockRender_posts($attr){
       $queryArgs['tag_id'] = $tagsToQuery;
     endif;
   endif;
-
   $filter_query = new WP_Query( $queryArgs );
-  // build content
+  // output
+  $output = '';
   if ( $filter_query->have_posts() ) :
-    $output .= '<div' . $id . ' class="block-posts' . $css . '" data-columnspace="' . $spacing . '" data-columns="' . $columns . '" style="' . $inlinecss . '">';
-      $output .= '<ul>';
-        while ( $filter_query->have_posts() ) : $filter_query->the_post();
-          $output .= '<li data-id="' . get_the_ID() . '">';
-            $get_url = get_post_meta(get_the_ID(), 'BlockUrl', true) ? get_post_meta(get_the_ID(), 'BlockUrl', true) : get_the_permalink();
-            $linkOpen = $get_url && $get_url !== '' ? '<a href="' . $get_url . '">' : '';
-            $linkClose = $get_url && $get_url !== '' ? '</a>' : '';
-            if(array_key_exists('postTaxonomyFilterOptions', $attr) && in_array('link_box', $attr['postTaxonomyFilterOptions'])):
-              $output .= $linkOpen;
+    while ( $filter_query->have_posts() ) : $filter_query->the_post();
+      $output .= '<li data-id="' . get_the_ID() . '">';
+        $get_url = get_post_meta(get_the_ID(), 'BlockUrl', true) ? get_post_meta(get_the_ID(), 'BlockUrl', true) : get_the_permalink();
+        $linkOpen = $get_url && $get_url !== '' ? '<a href="' . $get_url . '">' : '';
+        $linkClose = $get_url && $get_url !== '' ? '</a>' : '';
+        if(array_key_exists('postTaxonomyFilterOptions', $attr) && in_array('link_box', $attr['postTaxonomyFilterOptions'])):
+          $output .= $linkOpen;
+        endif;
+          // add post thumbnail
+          if(array_key_exists('postTaxonomyFilterOptions', $attr) && in_array('link_img', $attr['postTaxonomyFilterOptions']) && !in_array('link_box', $attr['postTaxonomyFilterOptions'])):
+            $output .= $linkOpen;
+          endif;
+            if(array_key_exists('postThumb', $attr) && $attr['postThumb'] !== false):
+            if($attr['postType'] == 'attachment'):
+              $mineType = get_post_mime_type();
+              $audioType = ['audio/mpeg3', 'audio/x-mpeg-3', 'video/mpeg', 'video/x-mpeg', 'audio/m4a', 'audio/ogg', 'audio/wav', 'audio/x-wav', 'audio/mpeg'];
+              $videoType = ['video/mp4', 'video/x-m4v', 'video/quicktime', 'video/x-ms-asf', 'video/x-ms-wmv', 'application/x-troff-msvideo', 'video/avi', 'video/msvideo', 'video/x-msvideo', 'video/ogg', 'video/3gpp', 'audio/3gpp', 'video/3gpp2', 'audio/3gpp2', 'video/mpeg'];
+              if(in_array($mineType, $audioType)):
+                $file = '<audio controls src="' . wp_get_attachment_url(get_the_id()) . '" data-id="' . get_the_id() . '" />';
+              elseif (in_array($mineType, $videoType)):
+                $file = '<video controls src="' . wp_get_attachment_url(get_the_id()) . '" data-id="' . get_the_id() . '" />';
+              else:
+                $file = wp_get_attachment_image(get_the_id(), 'full', true, array("data-id" => get_the_id()));
+              endif;
+              $output .= '<figure>' . $file . '</figure>';
+            else:
+              $output .= get_the_post_thumbnail() ? '<figure>' . get_the_post_thumbnail(get_the_id(), 'full', array("data-id" => get_the_id())) . '</figure>' : '';
             endif;
-              // add post thumbnail
-              if(array_key_exists('postTaxonomyFilterOptions', $attr) && in_array('link_img', $attr['postTaxonomyFilterOptions']) && !in_array('link_box', $attr['postTaxonomyFilterOptions'])):
-                $output .= $linkOpen;
-              endif;
-                if(array_key_exists('postThumb', $attr) && $attr['postThumb'] !== false):
-                if($attr['postType'] == 'attachment'):
-                  $mineType = get_post_mime_type();
-                  $audioType = ['audio/mpeg3', 'audio/x-mpeg-3', 'video/mpeg', 'video/x-mpeg', 'audio/m4a', 'audio/ogg', 'audio/wav', 'audio/x-wav', 'audio/mpeg'];
-                  $videoType = ['video/mp4', 'video/x-m4v', 'video/quicktime', 'video/x-ms-asf', 'video/x-ms-wmv', 'application/x-troff-msvideo', 'video/avi', 'video/msvideo', 'video/x-msvideo', 'video/ogg', 'video/3gpp', 'audio/3gpp', 'video/3gpp2', 'audio/3gpp2', 'video/mpeg'];
-                  if(in_array($mineType, $audioType)):
-                    $file = '<audio controls src="' . wp_get_attachment_url(get_the_id()) . '" data-id="' . get_the_id() . '" />';
-                  elseif (in_array($mineType, $videoType)):
-                    $file = '<video controls src="' . wp_get_attachment_url(get_the_id()) . '" data-id="' . get_the_id() . '" />';
-                  else:
-                    $file = wp_get_attachment_image(get_the_id(), 'full', true, array("data-id" => get_the_id()));
-                  endif;
-                  $output .= '<figure>' . $file . '</figure>';
-                else:
-                  $output .= get_the_post_thumbnail() ? '<figure>' . get_the_post_thumbnail(get_the_id(), 'full', array("data-id" => get_the_id())) . '</figure>' : '';
+            endif;
+          if(array_key_exists('postTaxonomyFilterOptions', $attr) && in_array('link_img', $attr['postTaxonomyFilterOptions']) && !in_array('link_box', $attr['postTaxonomyFilterOptions'])):
+            $output .= $linkClose;
+          endif;
+          // add content
+          $output .= '<div class="post-content">';
+            if(array_key_exists('postTextOne', $attr) && $attr['postTextOne'] !== ''):
+              $output .= $attr['postTextOne'] == 'title' ? '<h4>' : '<div>';
+                if(array_key_exists('postTaxonomyFilterOptions', $attr) && in_array('link_row1', $attr['postTaxonomyFilterOptions']) && !in_array('link_box', $attr['postTaxonomyFilterOptions'])):
+                  $output .= $linkOpen;
                 endif;
+                  $output .= WPgutenberg_block_postsValue($attr['postTextOne'], get_the_ID());
+                if(array_key_exists('postTaxonomyFilterOptions', $attr) && in_array('link_row1', $attr['postTaxonomyFilterOptions']) && !in_array('link_box', $attr['postTaxonomyFilterOptions'])):
+                  $output .= $linkClose;
                 endif;
-              if(array_key_exists('postTaxonomyFilterOptions', $attr) && in_array('link_img', $attr['postTaxonomyFilterOptions']) && !in_array('link_box', $attr['postTaxonomyFilterOptions'])):
-                $output .= $linkClose;
-              endif;
-              // add content
-              $output .= '<div class="post-content">';
-                if(array_key_exists('postTextOne', $attr) && $attr['postTextOne'] !== ''):
-                  $output .= $attr['postTextOne'] == 'title' ? '<h4>' : '<div>';
-                    if(array_key_exists('postTaxonomyFilterOptions', $attr) && in_array('link_row1', $attr['postTaxonomyFilterOptions']) && !in_array('link_box', $attr['postTaxonomyFilterOptions'])):
-                      $output .= $linkOpen;
-                    endif;
-                      $output .= WPgutenberg_block_postsValue($attr['postTextOne'], get_the_ID());
-                    if(array_key_exists('postTaxonomyFilterOptions', $attr) && in_array('link_row1', $attr['postTaxonomyFilterOptions']) && !in_array('link_box', $attr['postTaxonomyFilterOptions'])):
-                      $output .= $linkClose;
-                    endif;
-                  $output .= $attr['postTextOne'] == 'title' ? '</h4>' : '</div>';
+              $output .= $attr['postTextOne'] == 'title' ? '</h4>' : '</div>';
+            endif;
+            if(array_key_exists('postTextTwo', $attr) && $attr['postTextTwo'] !== ''):
+              $output .= $attr['postTextTwo'] == 'title' ? '<h4>' : '<div>';
+                if(array_key_exists('postTaxonomyFilterOptions', $attr) && in_array('link_row2', $attr['postTaxonomyFilterOptions']) && !in_array('link_box', $attr['postTaxonomyFilterOptions'])):
+                  $output .= $linkOpen;
                 endif;
-                if(array_key_exists('postTextTwo', $attr) && $attr['postTextTwo'] !== ''):
-                  $output .= $attr['postTextTwo'] == 'title' ? '<h4>' : '<div>';
-                    if(array_key_exists('postTaxonomyFilterOptions', $attr) && in_array('link_row2', $attr['postTaxonomyFilterOptions']) && !in_array('link_box', $attr['postTaxonomyFilterOptions'])):
-                      $output .= $linkOpen;
-                    endif;
-                      $output .= WPgutenberg_block_postsValue($attr['postTextTwo'], get_the_ID());
-                    if(array_key_exists('postTaxonomyFilterOptions', $attr) && in_array('link_row2', $attr['postTaxonomyFilterOptions']) && !in_array('link_box', $attr['postTaxonomyFilterOptions'])):
-                      $output .= $linkClose;
-                    endif;
-                  $output .= $attr['postTextTwo'] == 'title' ? '</h4>' : '</div>';
+                  $output .= WPgutenberg_block_postsValue($attr['postTextTwo'], get_the_ID());
+                if(array_key_exists('postTaxonomyFilterOptions', $attr) && in_array('link_row2', $attr['postTaxonomyFilterOptions']) && !in_array('link_box', $attr['postTaxonomyFilterOptions'])):
+                  $output .= $linkClose;
                 endif;
-              $output .= '</div>';
-              if(array_key_exists('postTaxonomyFilterOptions', $attr) && in_array('link_box', $attr['postTaxonomyFilterOptions'])):
-                $output .= $linkClose;
-              endif;
-          $output .= '</li>';
-        endwhile;
-        wp_reset_postdata();
-        // grid fixer
-        // if(array_key_exists('postSwiper', $attr) && $attr['postSwiper'] !== true && array_key_exists('postColumns', $attr) && $attr['postColumns'] > 1):
-        //   for ($x = 1; $x < $attr['postColumns']; $x++) {
-        //     $output .= '<span class="grid-fixer"></span>';
-        //   }
-        // endif;
-      $output .= '</ul>';
-    $output .= '</div>';
+              $output .= $attr['postTextTwo'] == 'title' ? '</h4>' : '</div>';
+            endif;
+          $output .= '</div>';
+          if(array_key_exists('postTaxonomyFilterOptions', $attr) && in_array('link_box', $attr['postTaxonomyFilterOptions'])):
+            $output .= $linkClose;
+          endif;
+      $output .= '</li>';
+    endwhile;
+    wp_reset_postdata();
+    // grid fixer
+    // if(array_key_exists('postSwiper', $attr) && $attr['postSwiper'] !== true && array_key_exists('postColumns', $attr) && $attr['postColumns'] > 1):
+    //   for ($x = 1; $x < $attr['postColumns']; $x++) {
+    //     $output .= '<span class="grid-fixer"></span>';
+    //   }
+    // endif;
+  else:
+    $output .= '<li class="wide"><p class="no-results">' . __( 'Sorry, but nothing matched your search terms. Please try again with some different keywords.', 'WPgutenberg' ) . '</p></li>';
   endif;
+  return $output;
+}
+
+// callback function
+function WPgutenberg_blockRender_posts($attr){
+  $output = '';
+  $css = '';
+  $inlinecss = '';
+  // add edvanced options (ID/CSS)
+  $id = array_key_exists('anchor', $attr) ? ' id="' . $attr['anchor'] . '"' : '';
+  $css .= array_key_exists('className', $attr) ? ' ' . $attr['className'] : '';
+  $css .= array_key_exists('align', $attr) ? ' align' . $attr['align'] : '';
+  // add swiper
+  $css .= array_key_exists('postSwiper', $attr) && $attr['postSwiper'] !== false ? ' gallery-swiper' : ' gallery-grid';
+  // add popup
+  $css .= array_key_exists('postPopUp', $attr) && $attr['postPopUp'] !== false ? ' add-popup' : '';
+  $css .= array_key_exists('postPopUpNav', $attr) && $attr['postPopUpNav'] !== false ? ' popup-preview' : '';
+  // reset spacing if only one column
+  $spacing = $attr['postColumnsSpace'];
+  // inline stylings
+  $columns = array_key_exists('postColumns', $attr) ? $attr['postColumns'] : 1;
+  $inlinecss .= '--postColumns: ' . $columns . ';';
+  $inlinecss .= ' --postColumnsSpace: ' . $spacing . 'px;';
+  // build content
+  $output .= '<div' . $id . ' class="block-posts' . $css . '" data-columnspace="' . $spacing . '" data-columns="' . $columns . '" style="' . $inlinecss . '">';
+    // sort options
+    if(array_key_exists('postSortNav', $attr) && $attr['postSortNav'] !== false):
+      $output .= '<div class="sort-options">';
+        if(array_key_exists('postSortNavOptions', $attr)):
+          foreach ($attr['postSortNavOptions'] as $key => $value) {
+            if (function_exists('acf_get_field')):
+              $term = acf_get_field($value);
+              if($term):
+                $name = $term['label'];
+              else:
+                $name = __( $value, 'WPgutenberg' );
+              endif;
+            else:
+              $name = __( $value, 'WPgutenberg' );
+            endif;
+            $sortd = $value == $attr['postSortBy'] ? $attr['postSortDirection'] : 'asc';
+            $sort_css = '';
+            $sort_css .= $value == $attr['postSortBy'] ? 'active' : "";
+            $sort_css .= $sortd == 'desc' ? ' z-a' : "";
+            $output .= '<label data-sort="' . $value . '" data-sortd="' . $sortd . '" tabindex="0" class="' . $sort_css . '">';
+              $output .= '<span class="sort-name">' . $name . '</span>';
+              $output .= '<span class="sort-arrow"><svg xmlns="http://www.w3.org/2000/svg" width="9.155" height="4.926" viewBox="0 0 9.155 4.926"><path d="M4.66 4.922a.88.88 0 0 0 .487-.208l3.676-3.158A.876.876 0 1 0 7.684.225l-3.11 2.667L1.464.225A.876.876 0 1 0 .328 1.556l3.68 3.154a.879.879 0 0 0 .652.208z" fill="#000"/></svg></span>';
+            $output .= '</label>';
+          }
+          // fields to resort
+          $output .= '<form class="thefilter">';
+            if(array_key_exists('postTaxonomyFilter', $attr)):
+              $output .= '<input type="hidden" name="postTaxonomyFilter" value="' . implode('__', $attr['postTaxonomyFilter']) . '">';
+            endif;
+            $output .= '<input type="hidden" name="postTaxonomyFilterRelation" value="' . $attr['postTaxonomyFilterRelation'] . '">';
+            $output .= '<input type="hidden" name="postType" value="' . $attr['postType'] . '">';
+            $output .= '<input type="hidden" name="postSortBy" value="' . $postSortBy . '">';
+            $output .= '<input type="hidden" name="postSortDirection" value="' . $attr['postSortDirection'] . '">';
+            if(array_key_exists('postTaxonomyFilterOptions', $attr)):
+              $output .= '<input type="hidden" name="postTaxonomyFilterOptions" value="' . implode('__', $attr['postTaxonomyFilterOptions']) . '">';
+            endif;
+            $output .= '<input type="hidden" name="postTextOne" value="' . $attr['postTextOne'] . '">';
+            $output .= '<input type="hidden" name="postTextTwo" value="' . $attr['postTextTwo'] . '">';
+            $output .= '<input type="hidden" name="postThumb" value="' . $attr['postThumb'] . '">';
+            $output .= '<input type="hidden" name="postColumns" value="' . $columns . '">';
+          $output .= '</form>';
+        endif;
+      $output .= '</div>';
+    endif;
+    // posts
+    $output .= '<ul>';
+      $output .= WPgutenberg_postresults_postssorting($attr);
+    $output .= '</ul>';
+  $output .= '</div>';
 
   return $output;
 }